// apps/integration_service/src/frameworks/support/interfaces/tokens-common.interface.ts
export interface ERC20TokenMinimal {
  address: string;
}

export interface ERC20TokenWithUnderlingMinimal extends ERC20TokenMinimal {
  underlying: ERC20TokenMinimal[];
}

// TODO: This needs to match asset service interface
export interface ERC20Token extends ERC20TokenMinimal {
  // ERC20 Standard
  address: string;
  symbol: string;
  name: string;
  decimals: number;
  totalSupply?: number;

  // Extra Token Info
  price: number;
  chainId: number;

  // User Details
  value?: number;
  balance?: number;

  // LP/wrapped/underlying token extras
  reserve?: number;
  position?: number;
  underlying?: ERC20Token[];
}

export interface ITokenMinimal<TExtra = void> {
  token: ERC20TokenMinimal; // just address
  extra?: TExtra;
}

export interface ITokenOpportunity<TExtra = void> extends ITokenMinimal<TExtra> {
  token: ERC20Token; // full erc20 token with price
}

export interface ITokenUserEntry<TExtra = void> extends ITokenOpportunity<TExtra> {
  amount: number; // user balance
  value: number; // balance * price
}

export interface IBaseApy {
  variableApy?: number;
  stableApy?: number;
}